requirements.txt
# --- Core API & Validation ---
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.19
pydantic==2.11.5
pydantic-settings==2.6.1

# --- PDF Processing (Docling) ---
docling==2.7.0
docling-core==2.7.0

# --- LlamaIndex Ecosystem ---
# Versions aligned for LlamaIndex 0.14.x
llama-index==0.14.13
llama-index-core==0.14.13
llama-index-vector-stores-qdrant==0.9.1
llama-index-embeddings-huggingface==0.6.1

# --- Vector Database ---
# BUMPED: Must be >= 1.16.0 to satisfy the vector store adapter
# BUMPED: Must be >= 1.16.0 to satisfy the vector store adapter
qdrant-client==1.16.1
fastembed==0.3.4

# --- Database (MongoDB) ---
pymongo==4.10.1
motor==3.7.1

# --- LLM Integration ---
groq==1.0.0

# --- ML & Image Processing ---
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3
pillow==10.4.0

# --- Utilities & Constraints ---
python-dotenv==1.0.1
aiofiles==23.2.1
# CRITICAL: Keep numpy pinned <2.0 for Docling compatibility
numpy==1.26.4
# sentence-transformers <3.0 is compatible with transformers 4.x (required for docling-ibm-models)
sentence-transformers>=2.0,<3.0